Provide solution for reverseString
Used a for loop to iterate over the length of
the input string param. Starting at the end
of the array and decrementing down, i put in each
character into an array to have it be reversed.

Comparing to the given solution, it was possible to
convert the inputString into an array, split it into
inidividual chars. .reverse the array order, then
.join the string back together. As I am learning
i did not know most of these methods existed.